♻️  Single source for non-error prefix
