+++ packages/core/src/domain/error/trackRuntimeError.ts
@@ -1,5 +1,6 @@
-import type { RawError } from '../../tools/error'
-import { ErrorSource, computeRawError, ErrorHandling } from '../../tools/error'
+import { NonErrorPrefix, ErrorSource, ErrorHandling } from '../../tools/error.types'
+import type { RawError } from '../../tools/error.types'
+import { computeRawError } from '../../tools/error'
 import type { Observable } from '../../tools/observable'
 import { clocksNow } from '../../tools/timeUtils'
 import { startUnhandledErrorCollection } from '../tracekit'
@@ -11,7 +12,7 @@ export function trackRuntimeError(errorObservable: Observable<RawError>) {
         stackTrace,
         originalError,
         startClocks: clocksNow(),
-        nonErrorPrefix: 'Uncaught',
+        nonErrorPrefix: NonErrorPrefix.UNCAUGHT,
         source: ErrorSource.SOURCE,
         handling: ErrorHandling.UNHANDLED,
       })
