+++ packages/core/src/tools/error.spec.ts
@@ -1,20 +1,14 @@
 import type { StackTrace } from '../domain/tracekit'
-import type { RawErrorCause, ErrorWithCause } from './error'
+import { NonErrorPrefix, ErrorSource, ErrorHandling } from './error.types'
+import type { RawErrorCause, ErrorWithCause } from './error.types'
+import { createHandlingStack, computeRawError, getFileFromStackTraceString, flattenErrorCauses } from './error'
 import { clocksNow } from './timeUtils'
-import {
-  createHandlingStack,
-  computeRawError,
-  getFileFromStackTraceString,
-  flattenErrorCauses,
-  ErrorSource,
-  ErrorHandling,
-} from './error'
 
 describe('computeRawError', () => {
   const NOT_COMPUTED_STACK_TRACE: StackTrace = { name: undefined, message: undefined, stack: [] } as any
   const DEFAULT_RAW_ERROR_PARMS = {
     startClocks: clocksNow(),
-    nonErrorPrefix: 'Uncaught',
+    nonErrorPrefix: NonErrorPrefix.UNCAUGHT,
     source: ErrorSource.CUSTOM,
   }
