+++ packages/core/src/tools/error.ts
@@ -1,48 +1,9 @@
 import type { StackTrace } from '../domain/tracekit'
 import { computeStackTrace } from '../domain/tracekit'
+import type { NonErrorPrefix, ErrorSource, ErrorHandling, RawError, ErrorWithCause, RawErrorCause } from './error.types'
 import { callMonitored } from './monitor'
 import type { ClocksState } from './timeUtils'
 import { jsonStringify, noop } from './utils'
-
-export interface ErrorWithCause extends Error {
-  cause?: Error
-}
-
-export type RawErrorCause = {
-  message: string
-  source: string
-  type?: string
-  stack?: string
-}
-
-export interface RawError {
-  startClocks: ClocksState
-  message: string
-  type?: string
-  stack?: string
-  source: ErrorSource
-  originalError?: unknown
-  handling?: ErrorHandling
-  handlingStack?: string
-  causes?: RawErrorCause[]
-}
-
-export const ErrorSource = {
-  AGENT: 'agent',
-  CONSOLE: 'console',
-  CUSTOM: 'custom',
-  LOGGER: 'logger',
-  NETWORK: 'network',
-  SOURCE: 'source',
-  REPORT: 'report',
-} as const
-
-export const enum ErrorHandling {
-  HANDLED = 'handled',
-  UNHANDLED = 'unhandled',
-}
-
-export type ErrorSource = typeof ErrorSource[keyof typeof ErrorSource]
 
 type RawErrorParams = {
   stackTrace?: StackTrace
@@ -50,7 +11,7 @@ type RawErrorParams = {
 
   handlingStack?: string
   startClocks: ClocksState
-  nonErrorPrefix: string
+  nonErrorPrefix: NonErrorPrefix
   source: ErrorSource
   handling: ErrorHandling
 }
