+++ packages/core/src/transport/sendWithRetryStrategy.ts
@@ -1,9 +1,9 @@
 import type { EndpointType } from '../domain/configuration'
 import { setTimeout } from '../browser/timer'
-import type { RawError } from '../tools/error'
+import { ErrorSource } from '../tools/error.types'
+import type { RawError } from '../tools/error.types'
 import { clocksNow } from '../tools/timeUtils'
 import { ONE_KIBI_BYTE, ONE_MEBI_BYTE, ONE_SECOND, ONE_MINUTE } from '../tools/utils'
-import { ErrorSource } from '../tools/error'
 import type { Payload, HttpResponse } from './httpRequest'
 
 export const MAX_ONGOING_BYTES_COUNT = 80 * ONE_KIBI_BYTE
